services/object/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.2
prometheus-client==0.17.0
python-dotenv==1.0.0
pyyaml==6.0
numpy==1.24.3
opencv-python-headless==4.7.0.72
torch==2.0.1
torchvision==0.15.2
tensorflow==2.12.0
darknet==0.3.4       # For YOLO
scikit-learn==1.2.2  # For additional ML utilities
